fix stack grow bug

+++ src/object.c
@@ -68,7 +68,7 @@ static Obj* allocateObject(uint64_t size, ObjType type) {
 }
 
 HOT_FUNCTION
-ObjUpvalue* newUpvalue(Value* slot, uint32_t offset)
+ObjUpvalue* newUpvalue(Value* slot, ptrdiff_t offset)
 {
 	ObjUpvalue* upvalue = ALLOCATE_OBJ(ObjUpvalue, OBJ_UPVALUE);
 	upvalue->location = slot;
